requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
webvtt-py==0.4.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.24.1
httpcore==0.17.3
